+++ decoders/sma_can/__init__.py
@@ -1,0 +1,5 @@
+'''
+SMA CAN protocol decoder.
+'''
+
+from .pd import Decoder